move envconfig
